requirements.txt
langchain
pinecone-client
python-dotenv
streamlit
pinecone
langchain-community